adding E2E Dose#1 scenario and some Renaming
