--- src/test/java/communityPortal/tests/Appointment_Scheduling/BookingDose1_Pneumo.java
@@ -1,4 +1,0 @@
-package communityPortal.tests.Appointment_Scheduling;
-
-public class BookingDose1_Pneumo {
-}
