--- src/test/java/communityPortal/tests/Appointment_Scheduling/BookingDose2_COVID19.java
@@ -1,4 +1,0 @@
-package communityPortal.tests.Appointment_Scheduling;
-
-public class BookingDose2_COVID19 {
-}
