--- src/test/java/communityPortal/tests/Appointment_Scheduling/Early_Booking_Appointment.java
@@ -1,4 +1,0 @@
-package communityPortal.tests.Appointment_Scheduling;
-
-public class Early_Booking_Appointment {
-}
