+++ src/test/java/communityPortal/tests/Appointment_Scheduling_CP/BookingDose1_Influenza.java
@@ -1,0 +1,4 @@
+package communityPortal.tests.Appointment_Scheduling_CP;
+
+public class BookingDose1_Influenza {
+}
