+++ src/test/java/communityPortal/tests/Appointment_Scheduling_CP/Early_Booking_Appointment.java
@@ -1,0 +1,4 @@
+package communityPortal.tests.Appointment_Scheduling_CP;
+
+public class Early_Booking_Appointment {
+}
