+++ src/test/java/communityPortal/tests/InventoryCP/E2E_Consumption_CP.java
@@ -1,4 +1,4 @@
 package communityPortal.tests.InventoryCP;
 
-public class E2E_Consumption {
+public class E2E_Consumption_CP {
 }
