--- src/test/java/communityPortal/tests/VaccineAdministration/E2E_Dose1_Influenza_Pneumo.java
@@ -1,4 +1,0 @@
-package communityPortal.tests.VaccineAdministration;
-
-public class E2E_Dose1_Influenza_Pneumo {
-}
