--- src/test/java/communityPortal/tests/VaccineAdministration/E2E_Dose2_Covid19.java
@@ -1,4 +1,0 @@
-package communityPortal.tests.VaccineAdministration;
-
-public class E2E_Dose2_Covid19 {
-}
