+++ src/test/java/communityPortal/tests/VaccineAdministration_CP/E2E_Dose1_Influenza_Pneumo_CP.java
@@ -1,0 +1,4 @@
+package communityPortal.tests.VaccineAdministration_CP;
+
+public class E2E_Dose1_Influenza_Pneumo_CP {
+}
